requirements.txt
torch
torchvision
opencv-python-headless
pyyaml
tqdm
numpy
